go, organise into dedicated directory and refactor
Refactor sensor.go into a package and module in the root of the repository.
