+++ .gitignore
@@ -1,3 +1,4 @@
 # Vim swap files
 **/*.swp
 **/__pycache__
+**/*.pyc
