+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/blinken/paradar
+
+require periph.io/x/periph v3.6.2+incompatible
